added specific intern pages and other fixes
