+++ src/components/RFAPosButton.tsx
@@ -2,30 +2,38 @@ import React from "react";
 import { Row } from "react-bootstrap";
 import { Link } from "react-router-dom";
 import { ColorThemes } from "@/colors";
+import {createUseStyles} from 'react-jss';
 
-export const RFAPosButton: React.FC<{ link: string; title: string }> = (
+export const RFAPosButton: React.FC<{ link: string; title: string, backgroundColor: string, backgroundHover: string }> = (
   props
 ) => {
-  const button = {
-    color: ColorThemes.black,
-    background: "#FA8D1F",
-    fontFamily: "Mazzard-H-Extrabold",
-    fontSize: "45px",
-    borderColor: ColorThemes.mainOrange,
-    borderRadius: "15px",
-    paddingTop: "20px",
-    paddingBottom: "20px",
-    margin: "auto",
-  };
+
+  const useStyles = createUseStyles({
+    button: (props:any) =>  ({
+      composes: "py-3 justify-content-between align-items-center",
+      fontFamily: "BeVietnam-ExtraBold",
+      transition: "200ms",
+      color: ColorThemes.black,
+      borderRadius: "15px",
+      backgroundColor: props.backgroundColor,
+      "&:hover": {
+        backgroundColor: props.backgroundHover,
+        transition: "200ms",
+      }
+    }),
+  })
+  
+  const classes = useStyles({backgroundColor: props.backgroundColor, backgroundHover: props.backgroundHover});
 
   return (
     <Link className="hyperlink" to={props.link}>
       <Row
-        className="pos-container-button justify-content-center align-items-center"
-        style={button}
+        className= {classes.button}
       >
         <h2 className="text-center">{props.title}</h2>
       </Row>
     </Link>
   );
 };
+
+
