+++ src/navbarroutes.ts
@@ -1,0 +1,116 @@
+export const enrollvolunteerabout = [
+    {
+      headerName: "Enroll",
+      to: "/enroll",
+      subPages: [
+        {
+          subPageName: "For Individual Learners",
+          to: "/enroll/individlearners",
+        },
+        {
+          subPageName: "For Schools",
+          to: "/enroll/schools",
+        },
+        {
+          subPageName: "Register",
+          to: "/enroll/register",
+        },
+      ]
+    },
+    {
+      headerName: "Volunteer",
+      to: "/volunteer",
+      subPages: [
+        {
+          subPageName: "Overview",
+          to: "/volunteer/overview",
+        },
+        {
+          subPageName: "Teachers Positions",
+          to: "/volunteer/teacherpos",
+        },
+        {
+          subPageName: "Intern Overview",
+          to: "/volunteer/internoverview",
+        },
+        {
+          subPageName: "Publicity Positions",
+          to: "/volunteer/publicitypositions",
+        },
+        {
+          subPageName: "Technical Positions",
+          to: "/volunteer/technicalpositions",
+        },
+        {
+          subPageName: "Business Positions",
+          to: "/volunteer/businesspositions",
+        },
+        {
+          subPageName: "Volunteer Application",
+          to: "/volunteer/volunteerapp",
+        },
+        {
+          subPageName: "Volunteer FAQs",
+          to: "/volunteer/volunteerfaq",
+        },
+      ]
+    },
+    {
+      headerName: "About",
+      to: "/about",
+      subPages: [
+        {
+          subPageName: "Overview",
+          to: "/about/overview",
+        },
+        {
+          subPageName: "Meet The Team",
+          to: "/about/meettheteam",
+        },
+        {
+          subPageName: "Affiliated Organizations",
+          to: "/about/afforgs",
+        },
+        {
+          subPageName: "Newsletter",
+          to: "/about/newsletter",
+        },
+      ]
+    }
+  ]
+export const contactaboutdonate = [
+  {
+    headerName: "Contact",
+    to: "/contact",
+    subPages: [
+      {
+        subPageName: "Contact",
+        to: "/contact",
+      },
+    ]
+  },
+  {
+    headerName: "News",
+    to: "/news",
+    subPages: [
+      {
+        subPageName: "News",
+        to: "/news",
+      },
+    ]
+  },
+  {
+    headerName: "Donate",
+    to: "/donate",
+    subPages: [
+      {
+        subPageName: "Donate",
+        to: "/donate",
+      },
+      {
+        subPageName: "Merch",
+        to: "https://www.bonfire.com/store/rfa/ ",
+      },
+    ]
+  }
+]
